Using the chain nonce, ignoring the transaction queue nonce for Parts/Acks transactions.
It is *not* guaranteed that nodes receiving the transactions are on the same block height as the sender. This can lead to transactions getting rejected by the transaction permission contract. Such transactions get removed on the receiving node, causing a Nonce gap preventing following re-tries to fail since they are sent with a larger nonce.